Added health bar to enemy, using a shader.
I can't believe a shader was the easiest way I could find to just show a bar of two colors like this.

Maybe we'll regret it and should actually use a SubViewport on a Sprite3D to a Control node, or even just a node that draws by hand, or something. Or a pair of drawn sprites so the bar can use a fancy PNG?